Starter code for match_history and player_history
Inserting players into player_history isn't working and I think it's because of the async issues with SQL. The match isn't inserted into match_history until the match is already over despite my calling it at the beginning because it happens async, so when I try to insert players into player_history at the beginning, this.match_id is undefined at that point, and gives an error because we declare that field to not be null.

I think this same issue is also preventing scores from being updated in match_history at the end of the game.